correction of spacing and image
