--- src/card-categorias-de-imoveis/card-category.css
@@ -1,16 +1,0 @@
-.card-container {
-  border-radius: 7px;
-  padding: 15px 26px 12px;
-}
-.card {
-  box-shadow: 0px 3px 4px rgba(153, 155, 168, 0.15);
-  cursor: pointer;
-  border-radius: 8px;
-  display: inline-block;
-}
-.img {
-  border-radius: 8px;
-}
-.card-text {
-  padding-left: 12px;
-}
